+++ docs/scripts/demos/theme/textfield.vue
@@ -1,13 +1,13 @@
 <template>
   <section class="component-section">
-    <h6 :class="$tt('headline6')">Textfield</h6>
+    <h6 :class="$tt('headline6')">Textfield (subtitle1)</h6>
     <div class="component-wrapper textfield-component">
       <ui-form-field>
         <ui-textfield v-model="value" helper-text-id="textfield-component-1">
           Label
         </ui-textfield>
         <ui-textfield-helper id="textfield-component-1" visible>
-          Assistive text
+          Assistive text (caption)
         </ui-textfield-helper>
       </ui-form-field>
 
@@ -19,7 +19,7 @@
           </template>
         </ui-textfield>
         <ui-textfield-helper id="textfield-component-2" visible>
-          Assistive text
+          Assistive text (caption)
         </ui-textfield-helper>
       </ui-form-field>
     </div>